agregar mascota a un grupo
