+++ src/test/kotlin/org/pets/history/controller/PetControllerTest.kt
@@ -142,43 +142,6 @@ class PetControllerTest : IntegrationTest() {
     }
 
     @Test
-    fun `Given the details of a pet, registers said pet and return it`() {
-        val pet = anyPet()
-
-        mockMvc.post("/pets") {
-            accept = MediaType.APPLICATION_JSON
-            contentType = MediaType.APPLICATION_JSON
-            content = mapper.writeValueAsString(pet)
-        }.andExpect {
-            status { isCreated() }
-            content { contentType(MediaType.APPLICATION_JSON) }
-            jsonPath("$.id") { isNumber() }
-            jsonPath("$.name") { value(pet.name) }
-            jsonPath("$.photo") { value(pet.photo) }
-            jsonPath("$.weight") { value(pet.weight) }
-            jsonPath("$.birthdate") { value(pet.birthdate.format(dateFormatter)) }
-            jsonPath("$.breed") { value(pet.breed) }
-            jsonPath("$.fur") { value(pet.fur) }
-            jsonPath("$.sex") { value(pet.sex.toString()) }
-            jsonPath("$.age") { value(pet.age) }
-        }
-    }
-
-    @Test
-    fun `Given the details of a pet, when said details contain a negative weight, fail to register the pet and return an error`() {
-        val pet = anyPet()
-        pet.weight = -1.0
-
-        mockMvc.post("/pets") {
-            accept = MediaType.APPLICATION_JSON
-            contentType = MediaType.APPLICATION_JSON
-            content = mapper.writeValueAsString(pet)
-        }.andExpect {
-            status { isBadRequest() }
-        }
-    }
-
-    @Test
     fun `Given a pet id that does exist and the details of a medical visit, registers said record and return it`() {
         val pet = petRepository.save(anyPet())
         val id = pet.id
